backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
openai
python-multipart
python-dotenv
requests